Fix bugs, add tests, validate events as they are received
