+++ setup.cfg
@@ -22,6 +22,7 @@ use_parentheses = True
 line_length = 88
 
 [mypy]
+plugins = pydantic.mypy
 follow_imports = skip
 ignore_missing_imports = true
 check_untyped_defs = true
