+++ zwave_js_server/model/controller/data_model.py
@@ -1,0 +1,35 @@
+"""Data model for a Z-Wave JS controller."""
+from typing import List
+
+from ...const import TYPING_EXTENSION_FOR_TYPEDDICT_REQUIRED
+from .statistics import ControllerStatisticsDataType
+
+if TYPING_EXTENSION_FOR_TYPEDDICT_REQUIRED:
+    from typing_extensions import TypedDict
+else:
+    from typing import TypedDict
+
+
+class ControllerDataType(TypedDict, total=False):
+    """Represent a controller data dict type."""
+
+    libraryVersion: str
+    type: int
+    homeId: int
+    ownNodeId: int
+    isSecondary: bool  # TODO: The following items are missing in the docs.
+    isUsingHomeIdFromOtherNetwork: bool
+    isSISPresent: bool
+    wasRealPrimary: bool
+    isStaticUpdateController: bool
+    isSlave: bool
+    serialApiVersion: str
+    manufacturerId: int
+    productType: int
+    productId: int
+    supportedFunctionTypes: List[int]
+    sucNodeId: int
+    supportsTimers: bool
+    isHealNetworkActive: bool
+    statistics: ControllerStatisticsDataType
+    inclusionState: int
